pyproject.toml
[project]
name = "hdc-algo"
requires-python = ">=3.10"
dynamic = ["version"]
description = "Algorithmic components for the Humanitarian Data Cube"
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}

classifiers = [
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Development Status :: 2 - Pre-Alpha",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Natural Language :: English",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    "dask[array]>=2024.10.0",
    "numba",
    "numpy>=2",
    "xarray",
]

[project.urls]
Issues = "https://github.com/WFP-VAM/hdc-algo/issues/"

[project.entry-points."numba_extensions"]
init = "hdc.algo.vendor.numba_scipy:_init_extension"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "hdc/algo/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["hdc"]

[project.optional-dependencies]
stats = [
    "scipy",
]
build = [
    "packaging>=25.0",
    "twine>=6.2.0",
    "wheel>=0.45.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pandas-stubs",
    "pytest>=9.0.2",
    "ruff>=0.14.8",
    "scipy-stubs",
]

[tool.ruff]
exclude = [
    "hdc/algo/vendor",
]
line-length = 105
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "D",      # pydocstyle
    "UP",     # pyupgrade
    "PL",     # pylint
    "RUF",    # ruff-specific rules
]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D203",   # 1 blank line required before class docstring (conflicts with D211)
    "D213",   # Multi-line docstring summary should start at the second line (conflicts with D212)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "D",      # Don't require docstrings in tests
    "PLR2004", # Magic value used in comparison
    "E741",   # Ambiguous variable name (e.g., `l` for lambda)
    "N803",   # Argument name should be lowercase
    "N806",   # Variable in function should be lowercase
    "PLR0912",  # Too many branches
    "PLR0913",  # Too many arguments
    "PLR0915",  # Too many statements
]
"hdc/algo/ops/**/*.py" = [
    "PLR0912",  # Too many branches (numba-optimized code)
    "PLR0913",  # Too many arguments (numba-optimized code)
    "PLR0915",  # Too many statements (numba-optimized code)
]
"hdc/algo/ops/autocorr.py" = ["N806"]  # Allow uppercase for mathematical notation (N, Sx, Sy, etc.)
"hdc/algo/ops/ws2dwcv.py" = ["N806"]  # Allow uppercase for mathematical notation (tr_H)
"hdc/algo/ops/ws2dwcvp.py" = ["N806"]  # Allow uppercase for mathematical notation (tr_H)
"hdc/algo/ops/whit.py" = ["N806"]  # Allow uppercase for mathematical notation (M)

[tool.ruff.lint.pydocstyle]
convention = "numpy"  # or "google" or "pep257" - adjust based on your preference

[tool.ruff.lint.pylint]
max-args = 10  # Maximum number of arguments in function definition
max-statements = 8  # Maximum number of statements in function

[tool.mypy]
exclude = [
    "hdc/algo/vendor",
]

[[tool.mypy.overrides]]
module = [
    "numba.*",
    "numba.core.types.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
